More mini refactor -- multiple slides to slides in slot
