+++ packages/11ty/_includes/components/lightbox/index.js
@@ -1,5 +1,5 @@
 const { html } = require('~lib/common-tags')
- 
+
 /**
  * Lightbox Tag
  * @todo add conditional rendering for epub and pdf when lightbox is included in `entry`
@@ -8,17 +8,18 @@ const { html } = require('~lib/common-tags')
  * @param      {Object}  globalData
  */
 module.exports = function (eleventyConfig, { page }) {
-  const lightboxSlides = eleventyConfig.getFilter('lightboxSlides')
+  const lightboxSlide = eleventyConfig.getFilter('lightboxSlide')
   const lightboxUI = eleventyConfig.getFilter('lightboxUI')
   const lightboxData = eleventyConfig.getFilter('lightboxData')
 
   return async function (figures=page.figures) {
     if (!figures) return
 
+    const slides = await Promise.all( figures.map( f => lightboxSlide(f) ) )
     return html`
       <q-lightbox>
         ${await lightboxData(figures)}
-        ${await lightboxSlides(figures)}
+        ${slides.join('')}
         ${lightboxUI(figures)}
       </q-lightbox>
     `
